+++ R/generics.R
@@ -11,7 +11,7 @@
 validate_alife <- function(x) {
   stopifnot(is.data.frame(x))
   my_cols <- c(## "event_type",
-               "lifetime", "fh", "uh",
+               "lifetime", "risk_set",
                "hazard", "se_log_hazard",
                "lower_ci", "upper_ci")
   stopifnot(all(my_cols %in% colnames(x)))
@@ -29,7 +29,7 @@ validate_alife <- function(x) {
 ##' @export
 ##' @examples
 ##' # Create a minimal data frame with the required columns
-##' df <- data.frame(lifetime = 1:2, fh = c(.2, .01), uh = c(.8, .20),
+##' df <- data.frame(lifetime = 1:2, risk_set = c(.8, .20),
 ##'                  hazard = c(0.1, 0.125), se_log_hazard = c(0.1, 0.1),
 ##'                  lower_ci = c(0.08, 0.1), upper_ci = c(0.12, 0.15))
 ##'
@@ -73,7 +73,7 @@ validate_acdf <- function(x) {
 ##' @export
 ##' @examples
 ##' # Create a minimal data frame with the required columns
-##' df <- data.frame(lifetime = 1:2, fh = c(.2, .01), uh = c(.8, .20),
+##' df <- data.frame(lifetime = 1:2, risk_set = c(.8, .20),
 ##'                  hazard = c(0.1, 0.125), se_log_hazard = c(0.1, 0.1),
 ##'                  lower_ci = c(0.08, 0.1), upper_ci = c(0.12, 0.15))
 ##'